Add files via upload
This has the simple civilisation simulation, using random action choice, with possible actions being gather food, increase population, and expand territory. There is a diminishing return on the food gathering for the number of people in a cell, so that expansion of territory is encouraged. There is also a death mechanic if food reaches 0, and a cell dies if population of the cell reaches 0 (though I am not 100% sure if it is working as intended. There is also information of the statuses of each cell, and tribe as a whole after each iteration (however would like to improve the layout in later editions). AI and analysis needs to be added.